+++ Rebus/Auditing/Sagas/SagaAuditingMetadataKeys.cs
@@ -1,0 +1,11 @@
+﻿namespace Rebus.Auditing.Sagas
+{
+    public class SagaAuditingMetadataKeys
+    {
+        public const string HandleQueue = "handlequeue";
+        
+        public const string SagaDataType = "sagadatatype";
+        
+        public const string SagaHandlerType = "sagahandlertype";
+    }
+}